changing DB datatype from Long to int to save some bytes

+++ src/main/java/redis/clients/jedis/BinaryJedis.java
@@ -2866,7 +2866,7 @@ public class BinaryJedis implements BasicCommands, BinaryJedisCommands, MultiKey
     }
   }
 
-  public Long getDB() {
+  public int getDB() {
     return client.getDB();
   }
